Moved tff_patch to root
